refactor(*): move to java 9
