+++ build.gradle
@@ -20,8 +20,8 @@ apply plugin: 'io.spring.dependency-management'
 group = 'com.github.lindenhoney'
 version = '1.1'
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = 1.9
+targetCompatibility = 1.9
 
 springBoot {
     buildInfo()
